Add a large handful of UI controls.
All these controls follow the conventions of jQuery UI's ThemeRoller system. These components are intended to be 100% look-and-feel compatible with any theme generated on the ThemeRoller web site, not to mention any other "skin" that abides by the same conventions.

So far, I've implemented all of jQuery UI's controls, except for the date picker (because date pickers are hard; WOE AND BLËRG). Other UI components,
like the in-place editor and auto-completer from scripty1, will require extending the ThemeRoller system. I'll be reaching out to the proper individuals to get that started.